+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add: tags hierarchy with color to better visualize tag relationships
 - add: source maps of the app
+- add: dataset information display in VariableInfo component
 - change: improve sorting logic for tag children in TagsListLevel component
 - refactor: consolidate app initialization logic and improve static rendering handling
 - refactor: externalize to svelte-fileapp for all URL and routing related code